+++ src/main/java/org/openpnp/machine/photon/PhotonFeeder.java
@@ -1,0 +1,485 @@
+package org.openpnp.machine.photon;
+
+import org.openpnp.ConfigurationListener;
+import org.openpnp.gui.support.Wizard;
+import org.openpnp.machine.photon.exceptions.FeedFailureException;
+import org.openpnp.machine.photon.exceptions.FeederHasNoLocationOffsetException;
+import org.openpnp.machine.photon.exceptions.NoSlotAddressException;
+import org.openpnp.machine.photon.exceptions.UnconfiguredSlotException;
+import org.openpnp.machine.photon.protocol.ErrorTypes;
+import org.openpnp.machine.photon.protocol.PhotonCommands;
+import org.openpnp.machine.photon.protocol.PacketResponse;
+import org.openpnp.machine.photon.sheets.FeederPropertySheet;
+import org.openpnp.machine.photon.sheets.SearchPropertySheet;
+import org.openpnp.machine.reference.ReferenceActuator;
+import org.openpnp.machine.reference.ReferenceFeeder;
+import org.openpnp.model.Configuration;
+import org.openpnp.model.Location;
+import org.openpnp.model.Solutions;
+import org.openpnp.spi.*;
+import org.simpleframework.xml.Attribute;
+import org.simpleframework.xml.Element;
+
+import javax.swing.*;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.function.IntConsumer;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import static org.openpnp.machine.photon.protocol.PhotonResponses.*;
+
+public class PhotonFeeder extends ReferenceFeeder {
+    public static final String ACTUATOR_DATA_NAME = "PhotonFeederData";
+    PhotonProperties photonProperties;
+
+    @Attribute(required = false)
+    protected String hardwareId;
+
+    protected Integer slotAddress = null;
+
+    @Attribute(required = false)
+    protected int partPitch = 4;
+
+    protected boolean initialized = false;
+
+    @Element(required = false)
+    private Location offset;
+
+    public PhotonFeeder() {
+        Configuration.get().addListener(new ConfigurationListener.Adapter() {
+            @Override
+            public void configurationLoaded(Configuration configuration) {
+                photonProperties = new PhotonProperties(configuration.getMachine());
+
+                // Ensure actuators are added to the machine when it has PhotonFeeders
+                getDataActuator();
+            }
+        });
+    }
+
+    @Override
+    public Location getPickLocation() throws Exception {
+        verifyFeederLocationIsFullyConfigured();
+
+        return offset.offsetWithRotationFrom(getSlot().getLocation());
+    }
+
+    private void verifyFeederLocationIsFullyConfigured() throws NoSlotAddressException,
+            UnconfiguredSlotException, FeederHasNoLocationOffsetException {
+        if(slotAddress == null) {
+            throw new NoSlotAddressException(
+                    String.format("Photon Feeder with address %s has no address. Is it inserted?", hardwareId)
+            );
+        }
+
+        if(getSlot().getLocation() == null) {
+            throw new UnconfiguredSlotException(
+                    String.format("The slot at address %s has no location configured.", slotAddress)
+            );
+        }
+
+        if(offset == null) {
+            throw new FeederHasNoLocationOffsetException(
+                    String.format("Photon Feeder with address %s has no location offset.", hardwareId)
+            );
+        }
+    }
+
+    public void setOffset(Location offsets) {
+        Object oldValue = this.offset;
+        this.offset = offsets;
+        firePropertyChange("offsets", oldValue, offsets);
+    }
+
+    public Location getOffset() {
+        return offset;
+    }
+
+    @Override
+    public void findIssues(Solutions solutions) {
+        super.findIssues(solutions);
+
+        if(hardwareId == null) {
+            return;
+        }
+
+        if(slotAddress != null && getSlot().getLocation() == null) {
+            solutions.add(new Solutions.PlainIssue(
+                    this,
+                    "Feeder slot has no configured location",
+                    "Select the feeder in the Feeders tab and make sure the slot has a set location",
+                    Solutions.Severity.Error,
+                    "https://github.com/openpnp/openpnp/wiki/Photon-Feeder#slots-and-feeder-locations"
+            ));
+        }
+
+        if(offset == null) {
+            solutions.add(new Solutions.PlainIssue(
+                    this,
+                    "Feeder has no configured offset",
+                    "Select the feeder in the Feeders tab and make sure the feeder has an offset location from the slot",
+                    Solutions.Severity.Error,
+                    "https://github.com/openpnp/openpnp/wiki/Photon-Feeder#slots-and-feeder-locations"
+            ));
+        }
+    }
+
+    @Override
+    public void prepareForJob(boolean visit) throws Exception {
+        for (int i = 0; i <= photonProperties.getFeederCommunicationMaxRetry(); i++) {
+            findSlotAddressIfNeeded();
+
+            initializeIfNeeded();
+
+            if (initialized) {
+                verifyFeederLocationIsFullyConfigured();
+
+                super.prepareForJob(visit);
+                return;
+            }
+        }
+
+        throw new Exception("Failed to find and initialize the feeder");
+    }
+
+    public void findSlotAddress() throws Exception {
+        findSlotAddress(true);
+    }
+
+    private void findSlotAddress(boolean force) throws Exception {
+        if (slotAddress != null && !force) {
+            return;
+        }
+
+        Actuator actuator = getDataActuator();
+        String feederAddressResponseString = actuator.read(PhotonCommands.getFeederAddress(hardwareId));
+
+        PacketResponse response = GetFeederAddress.decode(feederAddressResponseString);
+        if (!response.isOk()) {
+            ErrorTypes error = response.getError();
+
+            if (error == ErrorTypes.TIMEOUT) {
+                setSlotAddress(null);
+                return;
+            }
+        }
+        setSlotAddress(response.getFeederAddress());
+    }
+
+    private void findSlotAddressIfNeeded() throws Exception {
+        findSlotAddress(false);
+    }
+
+    private void initializeIfNeeded() throws Exception {
+        if (initialized || slotAddress == null) {
+            return;
+        }
+
+        Actuator actuator = getDataActuator();
+        String responseString = actuator.read(PhotonCommands.initializeFeeder(slotAddress, hardwareId));
+        PacketResponse response = InitializeFeeder.decode(responseString);
+
+        if (!response.isOk()) {
+            if (response.getError() == ErrorTypes.TIMEOUT) {
+                slotAddress = null;
+                return;
+            } else if (response.getError() == ErrorTypes.WRONG_FEEDER_UUID) {
+                PhotonFeeder otherFeeder = findByHardwareId(response.getUuid());
+                if (otherFeeder == null) {
+                    otherFeeder = new PhotonFeeder();
+                    otherFeeder.setHardwareId(response.getUuid());
+                    Configuration.get().getMachine().addFeeder(otherFeeder);
+                }
+
+                // This other feeder is in the slot we thought we were
+                otherFeeder.setSlotAddress(response.getFeederAddress());
+
+                return;
+            }
+        }
+
+        initialized = true;
+    }
+
+    static Actuator getDataActuator() {
+        Machine machine = Configuration.get().getMachine();
+
+        Actuator actuator = machine.getActuatorByName(ACTUATOR_DATA_NAME);
+
+        if (actuator == null) {
+            actuator = new ReferenceActuator();
+            actuator.setName(ACTUATOR_DATA_NAME);
+            try {
+                machine.addActuator(actuator);
+            } catch (Exception exception) {
+                exception.printStackTrace(); // TODO Probably need to log this, figure out why it can happen first
+            }
+        }
+
+        return actuator;
+    }
+
+    @Override
+    public void feed(Nozzle nozzle) throws Exception {
+        for (int i = 0; i <= photonProperties.getFeederCommunicationMaxRetry(); i++) {
+            findSlotAddressIfNeeded();
+            initializeIfNeeded();
+
+            if(! initialized) {
+                continue;
+            }
+
+            verifyFeederLocationIsFullyConfigured();
+
+            Actuator actuator = getDataActuator();
+            String ackResponseString = actuator.read(PhotonCommands.moveFeedForward(slotAddress, partPitch * 10));
+
+            PacketResponse ackResponse = MoveFeedForward.decode(ackResponseString);
+            if (!ackResponse.isOk()) {
+                slotAddress = null;
+                initialized = false;
+                ErrorTypes error = ackResponse.getError();
+                if (error == ErrorTypes.TIMEOUT) {
+                    throw new FeedFailureException("Feed command timed out");
+                } else if (error == ErrorTypes.UNINITIALIZED_FEEDER) {
+                    continue;
+                }
+            }
+
+            return;
+        }
+
+        throw new FeedFailureException("Failed to feed for an unknown reason. Is the feeder inserted?");
+    }
+
+    @Override
+    public String getPropertySheetHolderTitle() {
+        String classSimpleName = getClass().getSimpleName();
+        if(hardwareId == null) {
+            return String.format("Unconfigured %s", classSimpleName);
+        } else {
+            return String.format("%s %s", classSimpleName, getName());
+        }
+    }
+
+    @Override
+    public PropertySheet[] getPropertySheets() {
+        List<PropertySheet> sheets = new ArrayList<>();
+
+        if(hardwareId != null) {
+            sheets.add(new FeederPropertySheet(this));
+        }
+
+        sheets.add(new SearchPropertySheet());
+
+        return sheets.toArray(new PropertySheet[0]);
+    }
+
+    @Override
+    public PropertySheetHolder[] getChildPropertySheetHolders() {
+        return new PropertySheetHolder[0];
+    }
+
+    @Override
+    public Action[] getPropertySheetHolderActions() {
+        return new Action[0];
+    }
+
+    @Override
+    public Wizard getConfigurationWizard() {
+        return null;
+    }
+
+    @Override
+    public String getName() {
+        if(hardwareId == null) {
+            return String.format("Unconfigured %s", getClass().getSimpleName());
+        }
+
+        StringBuilder result = new StringBuilder();
+        result.append(name);
+        result.append(" (Slot: ");
+
+        if (slotAddress == null) {
+            result.append("None");
+        } else {
+            result.append(slotAddress);
+        }
+
+        result.append(")");
+
+        return result.toString();
+    }
+
+    @Override
+    public void setName(String name) {
+        Matcher matcher = Pattern.compile("(\\(Slot: [\\w+]+\\))").matcher(name);
+        while (matcher.find()) {
+            name = name.replace(matcher.group(), "");
+        }
+
+        name = name.trim();
+
+        super.setName(name);
+    }
+
+    /**
+     * The PhotonFeeder assumes you have a physical slot that is numbered 1 - 254. That
+     * value is also used in the protocol as the address of the feeder once the feeder
+     * is initialized.
+     *
+     * @return The slot address of this feeder or null if it doesn't have one.
+     */
+    public Integer getSlotAddress() {
+        return slotAddress;
+    }
+
+    public PhotonFeederSlots.Slot getSlot() {
+        if(slotAddress == null) {
+            return null;
+        }
+
+        return photonProperties.getFeederSlots().getSlot(slotAddress);
+    }
+
+    public void setSlotAddress(Integer slotAddress) {
+        PhotonFeederSlots.Slot oldSlot = getSlot();
+        Integer oldValue = this.slotAddress;
+        String oldName = this.getName();
+
+        if(slotAddress != null) {
+            // Find any other photon feeders and if they have this slot address, set their address to null
+            PhotonFeeder otherFeeder = findBySlotAddress(slotAddress);
+            if (otherFeeder != null) {
+                otherFeeder.slotAddress = null;
+                otherFeeder.initialized = false;
+            }
+        }
+
+        this.slotAddress = slotAddress;
+
+        firePropertyChange("slotAddress", oldValue, slotAddress);
+        firePropertyChange("slot", oldSlot, getSlot());
+        firePropertyChange("name", oldName, getName());
+    }
+
+    public String getHardwareId() {
+        return hardwareId;
+    }
+
+    public void setHardwareId(String hardwareId) {
+        String oldValue = this.hardwareId;
+        this.hardwareId = hardwareId;
+
+        if(getClass().getSimpleName().equals(name)) {
+            name = hardwareId;
+        }
+        
+        firePropertyChange("hardwareId", oldValue, hardwareId);
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return super.isEnabled() &&
+                hardwareId != null &&
+                partId != null &&
+                slotAddress != null &&
+                offset != null &&
+                getSlot().getLocation() != null;
+    }
+
+    public boolean isInitialized() {
+        return initialized;
+    }
+
+    public void setPartPitch(int partPitch) {
+        this.partPitch = partPitch;
+    }
+
+    public int getPartPitch() {
+        return partPitch;
+    }
+
+    public static PhotonFeeder findByHardwareId(String hardwareId) {
+        for (Feeder feeder : Configuration.get().getMachine().getFeeders()) {
+            if (!(feeder instanceof PhotonFeeder)) {
+                continue;
+            }
+
+            PhotonFeeder photonFeeder = (PhotonFeeder) feeder;
+
+            // Are we explicitly asking for feeders with null hardware Id?
+            if (photonFeeder.hardwareId == null && hardwareId == null) {
+                return photonFeeder;
+            }
+
+            if (photonFeeder.hardwareId != null && photonFeeder.hardwareId.equals(hardwareId)) {
+                return photonFeeder;
+            }
+        }
+
+        return null;
+    }
+
+    public static PhotonFeeder findBySlotAddress(int slotAddress) {
+        for (Feeder feeder : Configuration.get().getMachine().getFeeders()) {
+            if (!(feeder instanceof PhotonFeeder)) {
+                continue;
+            }
+
+            PhotonFeeder photonFeeder = (PhotonFeeder) feeder;
+
+            if (photonFeeder.slotAddress != null && photonFeeder.slotAddress.equals(slotAddress)) {
+                return photonFeeder;
+            }
+        }
+
+        return null;
+    }
+
+    public static void findAllFeeders(IntConsumer progressUpdate) throws Exception {
+        Machine machine = Configuration.get().getMachine();
+        PhotonProperties photonProperties = new PhotonProperties(machine);
+        Actuator actuator = getDataActuator();
+        int maxFeederAddress = photonProperties.getMaxFeederAddress();
+
+        List<PhotonFeeder> feedersToAdd = new ArrayList<>();
+
+        for (int address = 1; address <= maxFeederAddress; address++) {
+            String command = PhotonCommands.getFeederId(address);
+            String response = actuator.read(command);
+            PacketResponse packetResponse = GetFeederId.decode(response);
+
+            if (progressUpdate != null) {
+                int progress = (address * 100) / maxFeederAddress;
+                progressUpdate.accept(progress);
+            }
+
+            if (packetResponse.isOk()) {
+                PhotonFeeder otherFeeder = findByHardwareId(packetResponse.getUuid());
+                if (otherFeeder == null) {
+                    // Try to find an existing feeder without a hardware id before making a new one
+                    otherFeeder = findByHardwareId(null);
+                    if (otherFeeder == null) {
+                        otherFeeder = new PhotonFeeder();
+                        feedersToAdd.add(otherFeeder);
+                    }
+                }
+                otherFeeder.setHardwareId(packetResponse.getUuid());
+                otherFeeder.setSlotAddress(address);
+            } else if (packetResponse.getError() == ErrorTypes.TIMEOUT) {
+                PhotonFeeder otherFeeder = findBySlotAddress(address);
+                if (otherFeeder != null) {
+                    otherFeeder.slotAddress = null;
+                    otherFeeder.initialized = false;
+                }
+            }
+        }
+
+        for (PhotonFeeder feeder : feedersToAdd) {
+            Configuration.get().getMachine().addFeeder(feeder);
+        }
+    }
+}
